+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [x] fix response header casing for both errors and successful response
 - [ ] fix response body handling
 - [x] fix logging middleware not displaying timestamp
+- [ ] content-length header not generated for 404 with json error
 
 ### Changed